fix: align Model Intelligence API field names and endpoints with frontend
- analysis.py: rename response fields to short form (tp/tn/fp/fn), add f1
  and accuracy to summary; fix endpoint URLs (/confidence-histogram,
  /attack-by-type); return structured bin objects for histogram
- models.py: fix inference stats endpoint (/inference-stats)
- ModelSection.tsx: fix field references (f1_score→f1, auc_roc→auc)
- types/models.ts: update ModelTestResults interface (f1_score→f1, auc_roc→auc)
- queries.py: cast inference_time_ms to float for PERCENTILE_CONT

+++ dashboard/app/backend/db/queries.py
@@ -403,8 +403,8 @@ async def get_inference_stats(pool):
     SELECT
         COUNT(*) AS count,
         MIN(inference_time_ms) AS min_ms,
-        PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY inference_time_ms) AS p50_ms,
-        PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY inference_time_ms) AS p95_ms,
+        PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY inference_time_ms::float) AS p50_ms,
+        PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY inference_time_ms::float) AS p95_ms,
         MAX(inference_time_ms) AS max_ms,
         ROUND(AVG(inference_time_ms)::numeric, 2) AS avg_ms
     FROM gcn_predictions
